@@ -1,14 +1,15 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
-const Chatbox = (props) => {
+const Chatbox = props => {
   return (
     <div className="chatbox">
-     <p>{props.msg.message}</p>
+      <p>{props.msg.message}</p>
     </div>
   );
 };
 Chatbox.propTypes = {
-  msg: React.PropTypes.object.isRequired
+  msg: PropTypes.object.isRequired,
 };
 
 export default Chatbox;
