@@ -1,19 +1,20 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import Chatbox from './chatbox.js';
 
 // box man has two viewports (left and right), one for each eye
-const Viewport = (props) => {
+const Viewport = props => {
   return (
     <div id={props.id} className="viewport">
-      <video className="currentVideo" autoPlay></video>
-      <Chatbox msg={props.displayMsg}/>
+      <video className="currentVideo" autoPlay />
+      <Chatbox msg={props.displayMsg} />
     </div>
   );
 };
 Viewport.propTypes = {
-  id: React.PropTypes.string.isRequired,
-  displayMsg: React.PropTypes.object.isRequired
+  id: PropTypes.string.isRequired,
+  displayMsg: PropTypes.object.isRequired,
 };
 
 export default Viewport;
