@@ -1,3 +1,11 @@
+/**
+ * Borrowed some (not all) conventions from Scott Nonnenberg. 👍
+ * https://blog.scottnonnenberg.com/eslint-part-1-exploration/
+ *
+ * - Always use ‘error’ or ‘off’ instead of 0 and 2. Numbers are for real config values.
+ * - Rules are in alphabetical order: first core ESLint, then plugins sorted by name.
+ */
+
 module.exports = {
   env: {
     browser: true,
@@ -10,17 +18,30 @@ module.exports = {
     ecmaFeatures: {
       experimentalObjectRestSpread: true,
     },
-    ecmaVersion: 6,
+    // ecmaVersion: 6,
     sourceType: 'module',
   },
-  plugins: ['import', 'redux-saga', 'react', 'jsx-a11y'],
+  plugins: ['import', 'react', 'jsx-a11y'],
   globals: {
     NODE_ENV: true,
+    SERVER: true,
     describe: true,
     it: true,
     expect: true,
   },
   rules: {
+    'arrow-parens': 'off',
+    camelcase: 'off',
+    'comma-dangle': 'off',
+    'max-len': 'off',
+    'one-var': 'off',
+    'no-bitwise': 'off',
+    'no-case-declarations': 'off',
+    'no-console': 'off',
+    'no-debugger': 'off',
+    'no-nested-ternary': 'off',
+    'no-plusplus': 'off',
+    'no-unused-vars': 'error',
     indent: [
       'error',
       2,
@@ -28,29 +49,23 @@ module.exports = {
         SwitchCase: 1,
       },
     ],
-    // 'import/external-module-folders': ['node_modules'],
-    'import/no-extraneous-dependencies': 0,
-    'import/no-unresolved': 0,
-    'import/extensions': 0,
-    'import/first': 2,
-    // 'import/order': [
-    //   'error',
-    //   {
-    //     groups: ['builtin', 'external', 'parent', 'sibling', 'index'],
-    //     'newlines-between': 'always',
-    //   },
-    // ],
-    'react/jsx-filename-extension': 0,
-    'react/require-extension': 0,
-    'react/no-unescaped-entities': 0,
-    'linebreak-style': 0, //['error', 'unix'],
-    quotes: ['error', 'single'],
-    semi: ['error', 'always'],
-    'no-unused-vars': 0,
-    'max-len': 0,
-    'no-trailing-spaces': 0,
-    'comma-dangle': 0, // let prettier deal with that ish
-    'arrow-parens': 0,
-    'eol-last': 0,
+    'import/no-extraneous-dependencies': 'off',
+    'import/extensions': 'off',
+    'import/prefer-default-export': 'off',
+    'import/no-unresolved': 'off',
+    'jsx-a11y/accessible-emoji': 'off',
+    'jsx-a11y/href-no-hash': 'off',
+    'jsx-a11y/label-has-for': 'off',
+    'react/jsx-filename-extension': 'off',
+    'react/no-multi-comp': 'off',
+    'react/prop-types': 'off',
+    'react/sort-comp': 'off',
+  },
+  settings: {
+    'import/resolver': {
+      webpack: {
+        config: './webpack.common.js',
+      },
+    },
   },
 };
