@@ -4,18 +4,20 @@ import PropTypes from 'prop-types';
 
 import CharacterSelect from 'components/CharacterSelect.js';
 import CameraSelect from 'components/CameraSelect.js';
-import RTC from 'containers/RTC.js';
 import Game from 'components/Game.js';
+import Debug from 'components/Debug';
 import {
-  selectCharacter,
-  toggleFX,
-  changePhase,
-  changeGameroom,
-  changeConstraints,
   changeCamera,
+  changeConstraints,
+  changeGameroom,
+  changePhase,
+  enumerateCameras,
   foundCamera,
   fetchIceServers,
   initiateRTC,
+  selectCamera,
+  selectCharacter,
+  toggleFX,
 } from 'actions';
 
 class Application extends Component {
@@ -30,16 +32,9 @@ class Application extends Component {
     this.state = { localStream: null, remoteStreams: [] };
   }
 
-  componentDidMount() {
+  async componentDidMount() {
     /* Start 👁looking👁 for cameras as soon as component mounts */
-    navigator.mediaDevices
-      .enumerateDevices()
-      .then(deviceInfos => {
-        this.gotDevices(deviceInfos);
-      })
-      .catch(err => {
-        console.error(err);
-      });
+    this.props.enumerateCameras();
 
     // start up rtc stuff
     this.props.fetchIceServers();
@@ -54,55 +49,67 @@ class Application extends Component {
       console.log(window.remoteStreams);
       this.setState({ remoteStreams: window.remoteStreams });
     }
+
+    if (nextProps.streamChange.flag !== this.props.streamChange.flag) {
+      console.log('should handle stream change');
+    }
   }
 
   changeStream(stream, localOrRemote) {
     this.setState({ [`${localOrRemote}Stream`]: stream });
   }
 
-  gotDevices(deviceInfos) {
-    for (let i = 0; i < deviceInfos.length; i++) {
-      if (deviceInfos[i].kind === 'videoinput') {
-        this.props.foundCamera(deviceInfos[i]);
-      }
-    }
-  }
-
   render() {
     return (
       <div className="application">
-
+        <Debug>
+          <div>
+            localStream:{' '}
+            {JSON.stringify(
+              this.state.localStream && this.state.localStream.id
+            )}
+          </div>
+          <ul>
+            remoteStreams:{' '}
+            {this.state.remoteStreams.map(stream => (
+              <li>{stream && stream.id}</li>
+            ))}
+          </ul>
+        </Debug>
         <div id="background-title">
           <p id="title">BOX MAN</p>
           <p id="subtitle">OUT OF BODY</p>
         </div>
-        {this.props.phase === 'characterSelect'
-          ? <CharacterSelect
-              selectCharacter={this.props.selectCharacter}
-              toggleFX={this.props.toggleFX}
-              changePhase={this.props.changePhase}
-              changeGameroom={this.props.changeGameroom}
-            />
-          : this.props.phase === 'cameraSelect'
-              ? <CameraSelect
-                  character={this.props.character}
-                  availableCameras={this.props.availableCameras}
-                  changeStream={(stream, localOrRemote) => {
-                    this.changeStream(stream, localOrRemote);
-                  }}
-                  changePhase={this.props.changePhase}
-                  changeConstraints={this.props.changeConstraints}
-                />
-              : <Game
-                  localStream={this.state.localStream}
-                  remoteStreams={this.state.remoteStreams}
-                  fxMode={this.props.fxMode}
-                  constraints={this.props.constraints}
-                  character={this.props.character}
-                  gameroom={this.props.gameroom}
-                  realTimeConnection={this.props.realTimeConnection}
-                  messageFromPeer={this.props.messageFromPeer}
-                />}
+        {this.props.phase === 'characterSelect' ? (
+          <CharacterSelect
+            selectCharacter={this.props.selectCharacter}
+            toggleFX={this.props.toggleFX}
+            changePhase={this.props.changePhase}
+            changeGameroom={this.props.changeGameroom}
+          />
+        ) : this.props.phase === 'cameraSelect' ? (
+          <CameraSelect
+            character={this.props.character}
+            selectCamera={this.props.selectCamera}
+            availableCameras={this.props.availableCameras}
+            changeStream={(stream, localOrRemote) => {
+              this.changeStream(stream, localOrRemote);
+            }}
+            changePhase={this.props.changePhase}
+            changeConstraints={this.props.changeConstraints}
+          />
+        ) : (
+          <Game
+            localStream={this.state.localStream}
+            remoteStreams={this.state.remoteStreams}
+            fxMode={this.props.fxMode}
+            constraints={this.props.constraints}
+            character={this.props.character}
+            gameroom={this.props.gameroom}
+            realTimeConnection={this.props.realTimeConnection}
+            messageFromPeer={this.props.messageFromPeer}
+          />
+        )}
       </div>
     );
   }
@@ -118,6 +125,8 @@ const mapStateToProps = (state, ownProps) => ({
   realTimeConnection: state.realTimeConnection,
   fxMode: state.fxMode,
 
+  streamChange: state.streamChange,
+
   messageFromPeer: state.messageFromPeer,
 
   ...ownProps,
@@ -132,7 +141,10 @@ const mapDispatchToProps = dispatch => ({
   changeCamera: selectedCamera => dispatch(changeCamera(selectedCamera)),
   changeConstraints: constraints => dispatch(changeConstraints(constraints)),
 
+  enumerateCameras: () => dispatch(enumerateCameras()),
+
   foundCamera: newCamera => dispatch(foundCamera(newCamera)),
+  selectCamera: cameraInfo => dispatch(selectCamera(cameraInfo)),
 
   fetchIceServers: () => dispatch(fetchIceServers()),
   initiateRTC: () => dispatch(initiateRTC()),
